requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Deprecated==1.2.12
Django==2.2.4
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
idna==2.10
jwcrypto==0.9.1
oauthlib==3.1.1
pycparser==2.20
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
wrapt==1.12.1
